add checks before pet deletion

+++ petsoft/src/actions/actions.ts
@@ -95,6 +95,13 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 }
 
 export async function deletePet(petId: unknown) {
+  await sleep(1000);
+
+  const session = await auth();
+  if (!session?.user) {
+    redirect("/login");
+  }
+
   const validatedPetId = petIdSchema.safeParse(petId);
 
   if (!validatedPetId.success) {
@@ -104,7 +111,23 @@ export async function deletePet(petId: unknown) {
     };
   }
 
-  await sleep(1000);
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data,
+    }
+  })
+
+  if (!pet) {
+    return {
+      message: "Pet not found",
+    };
+  }
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "You do not have permission to delete this pet",
+    };
+  }
+ 
   try {
     await prisma.pet.delete({
       where: { id: validatedPetId.data },
